feat!: Add water bit flag to bitmap
This seems to be everything needed to make 2x2 bitmask work. This is
insane, with just 16 tiles we can generate every possible combination of
tiles. The trade off of using a dual grid system is miniscule. Apart
from the slighlty more complex backend (but significantly reduced
bitmasks), the only downside I see is that the rendered grid is shifted
by half the tile size (in my case 8 pixels). That's nothing, especially
in this game.