fix: Set GRASS_TYPE_INDEX if not water
Without this, the GRASS_TYPE_INDEX was never set when the tile was above
water level (with the exception of the paths).

This really illustrates how bad my current design is. I should
definitely spend some time to develop debugging tools and then see to it
that I refactor this mess.