feat: Add static YSort, fixes tree rendering issues
The problem was that my basic ysort function didn't work for
parent/child entities. Here the parent transform.translation.z affects
the child which is not accounted for.

This solution is very basic and if you have more than parent/child (e.g.
parent/child/grandchild) than you would need to define further structs.
I chose to go with this as it is the most simple to implement and
fairly fast (though not linear! I think). It does put the burden on the
user though.