fix: Prevent tree spawning at chunk edge
This gets rid of the issue that trees spawn on top of each other (side
by side). I think we only check that the trees are just not left or
bottom of the chunk because the position (v) only goes from 0 - 15 I
think, but it works so I will leave it for now.