fix: Return bird state to idle after picking an action
This fixes the weird bird glide, where the jumping animation didn't play
again. The issue was that the bird jumped, finished jumping and returned
to idling and then immediately went back to jumping because the player
was too close, all in the same frame (not changing the animations for at
least one frame).

Now the jumping animation finishes properly and restarts when we have
continues jumps.

This also explains why I was only able to see this behaviour sometimes,
but not all the time (because the order in which the functions were
executed changed every time the binary was built).