fix: Decrease bird YSort to render behind most objects
This kinda points out a bigger problem of YSort for me. It doesn't
exactly work to only have a single YSort value for all entities, at
least it seems that way.

The bird should have a lower YSort so that we can render it behind
Joanna, but too big and it will render behind bushes when it should be
rendered in front. I don't know what the solution is (partially because
I don't understand the problem good enough myself yet), but this
definitely isn't ideal.