feat!: Connect outer vertices path generation
This takes the graph created by Kruskals algorithm and adds loops to the
outer vertices (we determine outer vertices as only having one edge).

TODO: Currently the algorithm is pretty primitive and forces every
vertex in the graph to have at least two edges, even if it is really
hard to add a second one or doesn't make too much sense.

I didn't see any really bad cases yet but it's something to watch out
for.