fix!: Add Ending GameState, disable player input in ending
Implementing multiple states can get quite complex. Though it kinda is a
good way to validate whether your structure is good or bad. For example,
why on earth is the bird sound in `player/audio.mod`? That is super
unintuitive.

This also showed me that I should be more aware of states in general,
they are a powerful tool but you need to be able to use them properly.