feat: Add Ionas and Antonius NPCs
This was quite tricky as I had to accustom myself to having two NPCs
instead of just one. There were a few challenges with the current
implementation, but I didn't have to add or refactor too much. The
current implementation is also _okay_, it's gonna be good enough for
this game anyways.