fix!: Run update player animations BEFORE rapier writeback
This _hopefully_ fixes the animation glitch once and for all. The
problem was that we use the velocity of the player to update the state.
However the animation update could happen AFTER rapier did its
writeback, meaning the velocity could be almost zero, but not entirely
(because of some float inpercisions).

Not we update the animations before that happens, meaning we take
exactly the velocity that is being written by the player movement
function. Really goes to show how comparing floats to zero is just not a
good idea.