fix?: Start sakura pedal particles one frame after spawning them
This _might_ fix the issue of the first sakura pedals spawning in a
burst at (0, 0).

The problem is that I cannot reproduce the issue locally, neither on
debug build nor on wasm profile build. So I kinda have to just test in
prod, although the game isn't released anyways, so whatever I guess.