feat: Scale all UI based on the game screen size
This works pretty well, even works for the text. This allows the
developer (me) to just use `Val::Px` to hardcode the size and then
simply scale the UI to the corresponding screen size to make it work on
any size (doesn't mean it works on any resolution, only on 16:9, but I
don't really want to support others, sounds like a pain in the ass).

One problem with this is that Bevy is currently hanlding the font
atlases in a way that can _theoretically_ cause memory issues (running
out of memory). Given that the game window is not resizable and that in
browser you generally only have a fixed number of zoom levels, this
should not be _too_ problematic.

See `https://bevyengine.org/learn/errors/b0005/`.