fix: Check if birds are on water every frame
The birds were able to spawn on water and the way I implemented it
didn't catch that so the birds could just chill on the water for quite a
while if they didn't move (which is obviously shit and a bug).

Now we check every frame (every time we could perform an action) if the
bird is on water, if so, then fly somewhere else.