feat: Add more dialogue to Eleonore
Touches on #16. This should be fairly straightforward to implement, just
requires some time and care. I will also need to add some more
conditional branches when the player has already talked to the NPC they
are supposed to talk to (currently this is only implemented for
Antonions & Ionas dialogue).

There are also still some bugs with the dialogue graph.