restructure!: Remove Paladins entirely
It pains a bit, as quite a bit of time went into this, but I think this
is for the better. Keeping them would only be due to the sunk cost
fallacy. And fleshing out the dialogue would take a lot of time anyways.

I think this is another important lesson, if your game wants to be a
certain way, then you have to accept that. Even if you have spend a
dozen hours implementing a feature, if it just doesn't fit, remove it.

Less is often more.