restructure!: Remove fauna (birds) completely
Closes #9 and closes #13.

This is kinda painful as I spend a good 2-3 weeks on them, but I just
don't see them working really well unless I spend a huge amount of time
making the AI smart enough.

I don't consider it a waste of time though, I am sure that I will
eventually be able to implement birds in a smart way and once I am, I
can just take the birds implementation and make them work.

It was also a good showcase of how difficult even the smallest of AI's
are.